backend/requirements.txt
Django==2.2.19
djangorestframework-simplejwt==4.7.2
djangorestframework==3.13.1
djoser==2.1.0
Pillow==9.3.0
pytz==2022.5
sqlparse==0.4.3